requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
watchdog>=3.0.0
deepgram-sdk>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
sqlalchemy>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
aiofiles>=23.0.0
mutagen>=1.47.0
